+++ server/src/handler/eldenring/visit.rs
@@ -40,7 +40,7 @@ impl HandleRequest<Box<RequestGetVisitorListParams>, ResponseGetVisitorListParam
         let pool_matches = self.services.pool_visitor.matches(&VisitorPoolQuery {
             player_id: self.session.player_id,
             play_region: 0,
-            character_level: request.matching_parameters.character_level as u32,
+            character_level: request.matching_parameters.character_level,
             weapon_level: request.matching_parameters.max_reinforce as u32,
         });
